Make sure changes are seen
a) dialog is a bit more inefficient now, we no longer cache the ldap
data when either switching or initialising the computer/users view. This
probably can be solved by more a 'change' aware model
b) Also when returning from the propery sheet we make sure the current
item is selected.